@@ -1,24 +1,24 @@
 #include "vpkResourceProvider.h"
 
 #include <algorithm>
-#include <hook/steamAPI.h>
-#include <resource/resource.h>
-#include <resource/provider/filesystemResourceProvider.h>
-#include <utility/string/stringStrip.h>
+#include <hook/SteamAPI.h>
+#include <resource/Resource.h>
+#include <resource/provider/FilesystemResourceProvider.h>
+#include <utility/String.h>
 
 using namespace chira;
 
 VPKResourceProvider::VPKResourceProvider(unsigned int appID, const std::string& relativePath)
-    : AbstractResourceProvider(SOURCE_PROVIDER_NAME)
+    : IResourceProvider(SOURCE_PROVIDER_NAME)
     , absolutePath(calculateAbsolutePath(appID, relativePath))
     , vpk(this->absolutePath) {}
 
-bool VPKResourceProvider::hasResource(const std::string& name) const {
-    return !this->vpk.findEntry(name).fileName.empty();
+bool VPKResourceProvider::hasResource(std::string_view name) const {
+    return !this->vpk.findEntry(std::string{name}).fileName.empty();
 }
 
-void VPKResourceProvider::compileResource(const std::string& name, chira::Resource* resource) const {
-    auto entry = this->vpk.findEntry(name);
+void VPKResourceProvider::compileResource(std::string_view name, chira::Resource* resource) const {
+    auto entry = this->vpk.findEntry(std::string{name});
     std::vector<byte> output;
     this->vpk.readEntry(entry, output);
     output.push_back('\0');
@@ -28,19 +28,19 @@ void VPKResourceProvider::compileResource(const std::string& name, chira::Resour
 std::string VPKResourceProvider::calculateAbsolutePath(unsigned int appID, const std::string& relativePath) {
     std::string path = SteamAPI::Apps::getAppInstallPath(appID);
     FilesystemResourceProvider::nixifyPath(path);
-    stripRightModify(path, '/');
+    path = String::stripRight(path, '/');
     std::string relPath = relativePath;
     FilesystemResourceProvider::nixifyPath(relPath);
-    stripLeftModify(relPath, '/');
+    relPath = String::stripLeft(relPath, '/');
     return path + '/' + relPath;
 }
 
 FilesystemResourceProvider* VPKResourceProvider::getFilesystemResourceProvider(unsigned int appID, const std::string& relativePath) {
     std::string path = SteamAPI::Apps::getAppInstallPath(appID);
     FilesystemResourceProvider::nixifyPath(path);
-    stripRightModify(path, '/');
+    path = String::stripRight(path, '/');
     std::string relPath = relativePath;
     FilesystemResourceProvider::nixifyPath(relPath);
-    stripLeftModify(relPath, '/');
+    path = String::stripLeft(relPath, '/');
     return new FilesystemResourceProvider{path + '/' + relPath, true, SOURCE_PROVIDER_NAME};
 }
