@@ -1,6 +1,6 @@
 #pragma once
 
-#include <resource/provider/abstractResourceProvider.h>
+#include <resource/provider/IResourceProvider.h>
 #include <loader/vpk/vpk.h>
 
 const std::string SOURCE_PROVIDER_NAME = "source"; // NOLINT(cert-err58-cpp)
@@ -9,15 +9,15 @@ namespace chira {
     class FilesystemResourceProvider;
 }
 
-class VPKResourceProvider : public chira::AbstractResourceProvider {
+class VPKResourceProvider : public chira::IResourceProvider {
 public:
     explicit VPKResourceProvider(unsigned int appID, const std::string& relativePath);
-    [[nodiscard]] bool hasResource(const std::string& name) const override;
-    void compileResource(const std::string& name, chira::Resource* resource) const override;
+    [[nodiscard]] bool hasResource(std::string_view name) const override;
+    void compileResource(std::string_view name, chira::Resource* resource) const override;
     [[nodiscard]] std::string getPath() const {
         return this->absolutePath;
     }
-    const VPK& getVPK() const {
+    [[nodiscard]] const VPK& getVPK() const {
         return this->vpk;
     }
     [[nodiscard]] static std::string calculateAbsolutePath(unsigned int appID, const std::string& relativePath);
