@@ -1,6 +1,6 @@
 #pragma once
 
-#include <render/material/materialUntextured.h>
+#include <render/material/MaterialUntextured.h>
 #include "textureVTF.h"
 
 class MaterialVTF : public chira::MaterialUntextured {
@@ -11,5 +11,5 @@ public:
 protected:
     chira::SharedPointer<TextureVTF> vtfTex;
 private:
-    REGISTER_MATERIAL_TYPE(MaterialVTF);
+    CHIRA_REGISTER_MATERIAL_TYPE(MaterialVTF); // NOLINT(cert-err58-cpp)
 };
