@@ -1,13 +1,13 @@
 #pragma once
 
-#include <render/texture.h>
-#include <utility/logger.h>
+#include <core/Logger.h>
+#include <render/Texture.h>
 #include <resource/vtfTextureResource.h>
-#include <i18n/translationManager.h>
 
 class TextureVTF : public chira::Texture {
 public:
-    explicit TextureVTF(const std::string& identifier_, bool cacheTexture = true) : Texture(identifier_, cacheTexture) {}
+    explicit TextureVTF(const std::string& identifier_, bool cacheTexture = true)
+        : Texture(identifier_, cacheTexture) {}
     void compile(const nlohmann::json& properties) override;
     [[nodiscard]] unsigned int getWidth() const {
         return assert_cast<VTFImage*>(this->file->getFile())->getHeight();
